pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gtrend_api_tools"
version = "0.1.0_alpha"
description = "Tools for working with Google Trends APIs"
readme = "README.md"
authors = [
    {name = "David M. Perlman", email = "omgoleus+gtrend@gmail.com"}
]
requires-python = ">=3.6"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pandas",
    "requests",
    "pyyaml",
    "appdirs",
]

[project.optional-dependencies]
tor = ["stem"]
trendspy = ["trendspy"]
serpapi = ["google-search-results"]
all = ["stem", "trendspy", "google-search-results"]

[tool.setuptools]
packages = ["gtrend_api_tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v" 